requirements.txt
fastapi
openai
uvicorn
pydantic
python-dotenv
pymongo
langchain